Make exerslide work in Node v4
We were using array and parameter spread in some places, which is not
supported in Node v4 by default.
Interestingly, the error only occurred when actually running e.g.
`exerslide build`, but not when running the tests. The reason for that
is that we were using babel-register for the tests, which of course made
the code valid.
So now we split testing into two commands, one that tests the server
side code without `babel-register`, and one for client side code, with
the transformer.

The tests pass fine now in Node v4 and the example presentation builds
successfully.