[exerslide]: Handle props of wrapped component correctly in ExtensionPoint
The props of the wrapped component haven't been correctly made available
to the wrappers. The props need to be passed to the outer wrapper and
each wrapper has to pass them to their child (possibly augmented) (see
comments in the code).

Consider the following example:

<ExtensionPoint tags={['slide', 'a11y-announce-content', 'content']}>
  <div style={{fontColor: 'red'}}>
    {children}
  </div>
</ExtensionPoint>

If someone changed `js/SlideLayout.js` on a fresh installation to something
like the above (adding style to the wrapped element), that value would
be lost, i.e. it would be overridden by scaledContent.js, because
scaledContent.js doesn't get the correct props to merge the style into.

This commit fixes that.

I'm adding enzyme to test the behavior of core components and will be
adding more tests over time.