More helpful command line output
I'm taking a page from create-react-app here and suppress the default
output of the webpack-dev-server (since it is not useful or readable).

Instead, only a list of errors and warnings are printed, and they are
formatted the same for `exerslide build`, `exerslide watch` and
`exerslide serve`.

If there is a problem and you want to see the raw webpack output, pass
the `--verbose` options to any of the commands.

This also adds a new option to `exerslide serve`: `--no-open-browser`
which prevents exerslide from opening a browser window. Useful if you
don't want the browser to steal your focus.