Improved exerslide copy-defaults workflow (part 2)
Store hash of the template file in the copy.

This allows us to determine whether
 - the copy was manually edited and
 - whether the template file has been changed since the last time it was
   copied into the project.

With this we can avoid asking the user revealingly whether they want
to update a file.

Consider the scenario where the user customized the master layout.
Even if the master layout template hasn't been updated,
`exerslide copy-defaults` would detect the files as being different
and prompt the user to resolve the conflict. In this scenario, the
user would always want to to keep the file.
With the hash we can detect whether the template file changed and only
ask the user when it did.