[exerslide]: Fix layout CSS classes not being added to toc entries
This also adds a test for the `<TOC />` component. To get the test
working I had to `babel-plugin-module-alias` in order to resolve
references to `exerslide`, and `ignore-styles` to ignore the CSS file
imported by `TOC.js`.
Now however we should be able to test other components in the
scaffolding directory.