Prevent webpack babel loader from reading exerslide/.babelrc
Because of the way Babel works, it reads the .babelrc file if files
within `exerslide/` are transpiled. This is a problem because this file
is only supposed to be used for the browser unit tests.

There would be multiple ways to solve this, but the most self contained
seems to be to run the unit tests in a specific "babel environment" and
change the .babelrc file to only apply to that environment.