Fix a11y content visibility alert
The visibility alert hasn't been working for a while because the
document structure changed. That caused the a11y plugin to measure the
scroll and content height of the content itself, not the slide
"container". Because the content's height and scroll height is the same,
not alert was inserted.

The `a11y-announce-content` and `slide` extension point was moved to
wrap the slide container node, whose scroll and content height needs to
be measured instead.